Model PublishableConcern
